WIP add registration app
