+++ pyconkr/models.py
@@ -216,36 +216,3 @@ class Product(object):  # product is not django model now.
     def name(self):
         return 'PyCon Korea 2016'
 
-
-class Registration(models.Model):
-    user = models.ForeignKey(User)
-    merchant_uid = models.CharField(max_length=32)
-    name = models.CharField(max_length=100)
-    email = models.EmailField(max_length=255)
-    company = models.CharField(max_length=100, blank=True)
-    phone_number = models.CharField(max_length=20)
-    transaction_code = models.CharField(max_length=36)
-    payment_method = models.CharField(
-        max_length=20,
-        default='card',
-        choices=(
-            ('card', u'신용카드'),
-            #('vbank', _('Bank Transfer')),
-        )
-    )
-    payment_status = models.CharField(
-        max_length=10,
-        # choices=(
-        #     ('ready', _('Not Paid')),
-        #     ('failed', _('Failed')),
-        #     ('canceled', _('Canceled')),
-        #     ('paid', _('Paid')),
-        # )
-    )
-    payment_message = models.CharField(max_length=255, null=True)
-    vbank_num = models.CharField(max_length=255, null=True, blank=True)
-    vbank_name = models.CharField(max_length=20, null=True, blank=True)
-    vbank_date = models.CharField(max_length=50, null=True, blank=True)
-    vbank_holder = models.CharField(max_length=20, null=True, blank=True)
-    created = models.DateTimeField(auto_now_add=True)
-    modified = models.DateTimeField(auto_now=True)
