+++ pyconkr/urls.py
@@ -2,8 +2,7 @@ from django.conf import settings
 from django.conf.urls import patterns, include, url
 from django.contrib.flatpages import views
 from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-from .views import index, schedule, robots, registration_info, registration_payment, registration_status, \
-    registration_payment_callback
+from .views import index, schedule, robots
 from .views import RoomDetail
 from .views import AnnouncementList, AnnouncementDetail
 from .views import SpeakerList, SpeakerDetail, SpeakerUpdate
@@ -49,10 +48,11 @@ urlpatterns = patterns(
     url(r'^logout/$', logout, name='logout'),
     url(r'^profile/$', profile, name='profile'),
 
+    url(r'^registration/', include('registration.urls')),
     #url(r'^registration/$', registration_info, name='registration_info'),
-    url(r'^registration/payment/$', registration_payment, name='registration_payment'),
-    url(r'^registration/status/$', registration_status, name='registration_status'),
-    url(r'^registration/payment/callback/$', registration_payment_callback, name='registration_payment_callback'),
+    #url(r'^registration/payment/$', registration_payment, name='registration_payment'),
+    #url(r'^registration/status/$', registration_status, name='registration_status'),
+    #url(r'^registration/payment/callback/$', registration_payment_callback, name='registration_payment_callback'),
     url(r'^robots.txt$', robots, name='robots'),
     url(r'^summernote/', include('django_summernote.urls')),
     url(r'^admin/', include(admin.site.urls)),
