+++ registration/templates/registration/status.html
@@ -1,0 +1,94 @@
+{% extends "base.html" %}
+{% load i18n %}
+
+{% block content %}
+    <div>
+        <h3>등록확인</h3>
+        {% if registration.payment_status == 'paid' %}
+            <div>등록 완료</div>
+            <div>
+                <p>정상적으로 등록이 완료되었습니다.</p>
+            </div>
+            <table class="table table-striped table-bordered">
+                <tr>
+                    <th>이름</th>
+                    <td>{{ registration.name }}</td>
+                </tr>
+                <tr>
+                    <th>이메일</th>
+                    <td>{{ registration.email }}</td>
+                </tr>
+                <tr>
+                    <th>소속</th>
+                    <td>{{ registration.company }}</td>
+                </tr>
+                <tr>
+                    <th>전화번호</th>
+                    <td>{{ registration.phone_number }}</td>
+                </tr>
+                <tr>
+                    <th>금액</th>
+                    <td>15,000원</td>
+                </tr>
+                <tr>
+                    <th>거래번호</th>
+                    <td>{{ registration.merchant_uid }}</td>
+                </tr>
+                <tr>
+                    <th>등록완료일</th>
+                    <td>{{ registration.modified }}</td>
+                </tr>
+            </table>
+        {% elif registration.payment_status == 'ready' %}
+            <div>
+                <h3>결제정보</h3>
+                <table class="table table-striped table-bordered">
+                    <tr>
+                        <th>이름</th>
+                        <td>{{ registration.name }}</td>
+                    </tr>
+                    <tr>
+                        <th>이메일</th>
+                        <td>{{ registration.email }}</td>
+                    </tr>
+                    <tr>
+                        <th>가격</th>
+                        <td>15,000원</td>
+                    </tr>
+                    <tr>
+                        <th>거래번호</th>
+                        <td>{{ registration.merchant_uid }}</td>
+                    </tr>
+                    <tr>
+                        <th>입금하실 은행</th>
+                        <td>{{ registration.vbank_name }}</td>
+                    </tr>
+                    <tr>
+                        <th>입금하실 계좌번호</th>
+                        <td>{{ registration.vbank_num }}</td>
+                    </tr>
+                    <tr>
+                        <th>입금기한</th>
+                        <td>{{ registration.vbank_date }}</td>
+                    </tr>
+                    <tr>
+                        <th>입금확인</th>
+                        <td>
+                            {% if registration.payment_status == 'ready' %}
+                                입금확인중
+                            {% elif registration.payment_status == 'paid' %}
+                                입금확인완료
+                            {% endif %}
+                        </td>
+                    </tr>
+                </table>
+            </div>
+        {% else %}
+            <div>등록 내역이 없습니다.</div>
+            <div>
+                <p></p>
+                <a href="{% url 'registration_payment' %}">등록하기</a>
+            </div>
+        {% endif %}
+    </div>
+{% endblock %}
